Moves files to folders where they all should be
