+++ packages/client/src/components/measurables/MeasurablesSeriesTable.re
