+++ packages/client/src/components/notebooks/NotebookSidebar.re
